Create data_cleaning_revised.r
I propose the above revised code to clean the data. The step is the same, download a latest copy from OneDrive, save as UTF-8 csv file and run the code. Altogether 6 cloumns are added: "header"_1946_before, and "header"_1946_after. In short names they are denoted as SB and SA in the summary statistics. 

PS: The increase you can see in the summary stats of Sp between SB and SA should be related to tax intervals, some specific tax rates are removed and thus has a smaller denominator, increases the mean value.