Update data_cleaning.r
I propose to replace the previous step three with these new lines, the only difference is to add one more step of transformation within every f_function, the result should be the same. Also I confirmed with the the other two typos you mentioned, yes I agree with you, they are bad mistakes, I have updated them in the UnitsKeys in the excel file and very sorry for the inconvenience it brings.